Number of threads are now specified outside the Makefile (on main.c). Done some refactoring
The default number of threads is now obtanied via omp_get_max_threads, instead of being specified on Makefile, which was a pain. If compiler does not know the size of farr (which is the number of threads) will fail to build efficient code, so pointer must be avoided and using "TYPE farr[][SIZE]" works as intended.